fast-api/requirements.txt
# Core FastAPI dependencies
fastapi==0.115.3
uvicorn==0.25.0
pydantic
starlette==0.41.2
jinja2==3.1.4
python-multipart==0.0.18

# Database connections
psycopg2-binary==2.9.10
neo4j==5.28.1
pgvector==0.2.3

# Authentication and security
bcrypt==4.1.3
shortuuid==1.0.13

# NLP/ML core libraries 
sentence-transformers==2.7.0
huggingface_hub==0.23.4
transformers==4.38.1
scikit-learn==1.4.2
numpy==1.26.4
rouge_score==0.1.2
bert-score==0.3.13

# LangChain components
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.45
chromadb==0.6.3

# LLM integration
ollama==0.4.7

# Utilities
requests==2.32.3
tqdm==4.66.4
contextvars
asyncio
pathlib

# Telemetry (used in api_app.py)
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-sdk==1.29.0